+++ source/lib/AudioAnalyzer.h
@@ -57,7 +57,7 @@ namespace AudioVisualizer
 
 			long GetPosition() const { return _spInputBuffer->GetPosition(); }
 
-			UINT32 GetDownsampleRate() const { return _spInputBuffer->GetDownsampleFactor(); }
+			UINT32 GetDownsampleRate() const { return (UINT32) _spInputBuffer->GetDownsampleFactor(); }
 		};
 	}
 }
